requirements.txt
# Core dependencies for SAP Generative AI Hub SDK
sap-ai-sdk-gen[all]>=5.8.0

# Environment and configuration management
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# FastAPI and web dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# HTTP client for testing
httpx>=0.25.0

# CLI and utilities
click>=8.0.0
rich>=13.0.0

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
